requirements.txt
Flask==2.3.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.0
pandas==2.0.0
numpy==1.24.0
joblib==1.3.0
